% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{areas_gt}
\alias{areas_gt}
\title{Drop noncontiguous polygons with an area smaller or equal lower_bound.}
\usage{
areas_gt(an_area, lower_bound = 0)
}
\arguments{
\item{an_area}{A sp object, commonly a shapefile,  or a SDM_area object representing the area of study.}

\item{lower_bound}{A lower bound area indicating polygons which it going to dropped out.}
}
\value{
A sp object with remaining disaggregated polygons with area greater than lower_bound, or a SDM_area
containing a sp object with remaining disaggregated polygons with area greater than lower_bound.
Furthermore, if no coordinate reference system are used it will be set to EPSG:4326.
}
\description{
Drop noncontiguous polygons with an area smaller or equal lower_bound.
}
\examples{
\dontrun{

main <- cbind(
  c(0, 0, 1, 1),
  c(0, 1, 1, 0)
)
secondary <- cbind(
  c(1, 1.3, 1.3, 1),
  c(1, 1.0, 0.7, 0.7)
)
hole <- main / 3 + 1 / 3
island <- cbind(
  c(1.05, 1.05, 1.55, 1.55),
  c(0, .5, .5, 0)
)
P <- Polygons(
  ID = 1,
  list(
    Polygon(main),
    Polygon(hole, hole = TRUE),
    Polygon(island),
    Polygon(secondary)
  )
)

SP <- SpatialPolygons(list(P))

plot(SP)
}
}
