% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_nearest_land.R
\name{assign_nearest_land}
\alias{assign_nearest_land}
\title{Assign to nearest raster cell on mask}
\usage{
assign_nearest_land(dat_object, mask_object, max_distance, verbose = TRUE)
}
\arguments{
\item{dat_object}{data.frame}

\item{mask_object}{'SpatRaster'}

\item{max_distance}{map units if raster is projected}

\item{verbose}{provide verbose output. Default is TRUE}
}
\value{
modified data object minus observations not within specified maximum distance of mask
}
\description{
Adapted from seegSDM.
Reposition point observations to a location within mask if within a specified distance.
Useful for when coastal observations drop off jagged mask and similar
}
