% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mask.R
\name{example_mask}
\alias{example_mask}
\title{Create an example mask within a given raster}
\usage{
example_mask(raster, pc_threshold = NULL)
}
\arguments{
\item{raster}{generated from example_raster for example}

\item{pc_threshold}{A threshold percentile to divide mask / non mask elements}
}
\value{
A \code{spatRaster}
}
\description{
Create an example mask within a given raster
}
\examples{
example_mask(example_raster(), pc_threshold=0.5)
}
