% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_raster}
\alias{example_raster}
\title{Create example raster for use in examples}
\usage{
example_raster()
}
\value{
raster
}
\description{
Create example raster for use in examples
}
\examples{
example_raster
}
