% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Extract covariates for presence only data}
\usage{
extract_covariates(presences, absences, covariates)
}
\arguments{
\item{covariates}{}
}
\description{
Extracts
}
