% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitatAssociation.R
\name{habitatAssociation}
\alias{habitatAssociation}
\title{Infer habitat associations from a species sdm}
\usage{
habitatAssociation(raster, sdm, n = 10000, col = TRUE)
}
\arguments{
\item{raster}{A categorical SpatRaster.}

\item{sdm}{A SpatRaster with numerical values.}

\item{n}{Number of points to sample from the sdm.}

\item{col}{Logical. Use colours associated with the raster if any? Default \code{TRUE}. if \code{FALSE}, no colours used.}
}
\value{
A list with a ggplot object and a \code{data.frame} with the habitat association values.
}
\description{
This function quantifies habitat associations by contrasting what is available in an area to what is use by the species
}
\details{
This function compares the frequency of the different categories of habitats in the area to the frequencies of the habitats in the distribution of the species. First, it randomly samples points proportionally to the values in the sdm and calculates and
}
\examples{
\dontrun{

library(terra)
library(sf)
library(rstac)
library(exactextractr)
library(rstac)
library(ggplot2)
library(geodata)
library(rmapshaper)

# Downloads polygons using package geodata
can <- gadm("CAN", level = 1, path = getwd()) |> st_as_sf()
labrador <- ms_explode(can[can$NAME_1 \%in\% c("Newfoundland and Labrador"), ])
labrador <- labrador[which.max(st_area(labrador)), ] # keep Labarador
qc <- can[can$NAME_1 \%in\% c("Québec"),] |>
  rbind(labrador) |>
  ms_simplify(0.01)



# https://www.cec.org/north-american-environmental-atlas/land-cover-30m-2020/

category <- c("Temperate or sub-polar needleleaf forest",
             "Sub-polar taiga needleleaf forest",
             "Tropical or sub-tropical broadleaf evergreen forest",
             "Tropical or sub-tropical broadleaf deciduous forest",
             "Temperate or sub-polar broadleaf deciduous forest",
            "Mixed Forest",
             "Tropical or sub-tropical shrubland",
             "Temperate or sub-polar shrubland",
             "Tropical or sub-tropical grassland",
             "Temperate or sub-polar grassland",
            "Sub-polar or polar shrubland-lichen-moss",
             "Sub-polar or polar grassland-lichen-moss",
             "Sub-polar or polar barren-lichen-moss",
             "Wetland",
            "Cropland",
             "Barren lands",
             "Urban",
             "Water",
             "Snow and Ice")

dat <- data.frame(code = c(0, seq_along(category)), category = c("No data", category))

cec <- rast("/vsicurl/https://object-arbutus.cloud.computecanada.ca/bq-io/io/CEC_land_cover/NA_NALCMS_landcover_2020_30m.tif")

levels(cec) <- dat

cec <- crop(cec, st_transform(qc, st_crs(cec)))
raster <- aggregate(cec, 10, fun = "modal") |>
 crop(st_transform(qc, st_crs(cec)), mask = TRUE) |>
 droplevels()

io <- stac("https://acer.biodiversite-quebec.ca/stac/")


coll <- "oiseaux"
#coll <- "sdm_emv_finance"

urls <- io |>
  stac_search(collections = coll) |>
 post_request() |>
 items_fetch() |>
 _$features |>
 sapply(function(i){i$assets[[1]]$href})

sdm <- rast(paste0('/vsicurl/', grep("(?=.*1900-2024)(?=.*setophaga_cer)", urls, perl = TRUE, value = TRUE)))

x <- habitatAssociation(raster = raster, sdm = sdm, n = 10000, col = TRUE)
x

}

}
