% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rasts.R
\name{import_rasts}
\alias{import_rasts}
\title{Import rasters from directory}
\usage{
import_rasts(path, ext = ".grd", as_list = c(FALSE, TRUE))
}
\arguments{
\item{path}{Directory path containing rasters}

\item{ext}{Extension type}

\item{as_list}{Should the \code{spatRaster} objects be returned as a \code{list} (\code{TRUE}) or concatenated as layers in a single \code{spatRaster} object (\code{FALSE}; default)}
}
\value{
A \code{SpatRaster} if \code{as_list = FALSE}, or list of \code{SpatRaster} objects.
}
\description{
Imports all rasters of a given extension type from a specified directory using \code{terra::rast}.
Based on \code{seegSDM::importRasters}.
}
\examples{
\dontrun{
  rasters <- import_rasts("/data/grids/covariates")
}
}
