% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_africa_mask.R
\name{make_africa_mask}
\alias{make_africa_mask}
\title{Make Africa Mask}
\usage{
make_africa_mask(
  file_name = NULL,
  type = c("raster", "vector"),
  res = c("high", "low"),
  countries = NULL
)
}
\arguments{
\item{file_name}{Character of file path and name if  mask is to be written to disc.}

\item{type}{Character \code{raster} or \code{vector}; to return mask as either \code{SpatRaster} or \code{SpatVector}.}

\item{res}{Character \code{"high"} or \code{"low"}; corresponding to resolution of 0.008333333 or 0.04166667 decimal degrees}

\item{countries}{Character of ISO3 country names. If \code{NULL} returns all countries in Africa.}
}
\value{
\code{SpatRaster} or \code{SpatVector} in WGS 84 (EPSG:4326).
}
\description{
Makes a \code{spatRaster} mask layer of Africa, based on shapefiles for African nations from the \code{malaraAtlas} package. Function is a little slow, so writing to disc may be preferred.
}
\details{
Raster layers creates with extent of \code{terra::ext(-18.0000019073486, 52.0416647593181, -34.9999987284343, 37.5416679382324)}
}
\examples{
# Create an object in workspace
africa_mask_v <- make_africa_mask(type = "vector")

# Save to disk
make_africa_mask(file_name = "africa_mask.tif", type = "raster")

# or do both at once
africa_mask_r <- make_africa_mask("africa_mask.tif")
}
