% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mpp_list.R
\name{make_mpp_list}
\alias{make_mpp_list}
\title{Make presence-only list for \code{multispeciesPP}}
\usage{
make_mpp_list(x, id)
}
\arguments{
\item{x}{A \code{data.frame} containing covariate and bias values for locations,
and an identity (species) column}

\item{id}{The name of the identity column in \code{tidyselect} form. \code{id} must be
a column name in \code{x}.}
}
\value{
A named list or data frames.
}
\description{
Make presence-only list for \code{multispeciesPP}
}
\details{
The package \code{multispeciesPP} requires a very annoying named list
format for presence-only data. This function takes a table of covariate /
bias values and an ID column (i.e. species), and returns a named list where
each element is the values corresponding to that identity only. Ugh.
}
\examples{
\dontrun{
make_mpp_list(presence_only_data, species)
}
}
