% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_all.R
\name{mask_all}
\alias{mask_all}
\title{Mask all}
\usage{
mask_all(rasts, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{rasts}{\code{SpatRaster} with \code{nlyr(rasts)} > 1 to mask}

\item{filename}{\code{character}to save output}

\item{overwrite}{\code{logical} overwrite existing \code{filename}?}
}
\value{
\code{SpatRaster}
}
\description{
Masks all \code{NA} cells across all layers, such that returned layers
have matching \code{NA} cells.
}
\examples{

# Create some SpatRaster layers with non-matching NA cells
library(terra)
library(sdmtools)

r <- c(
  example_raster(seed = 1),
  example_raster(seed = 2),
  example_raster(seed = 3)
)
rvals <- terra::values(r)
nas <- c(1:10, 105:120, 215:240)
rvals[nas] <- NA
r[] <- rvals

# check if it pleases you to do so
# plot(r)
# mask out non-overlapping `NA` values in all layers
s <- mask_all(r)
s

# plot(s)
}
