% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_from_all.R
\name{mask_from_all}
\alias{mask_from_all}
\title{Create mask from raster layers}
\usage{
mask_from_all(r)
}
\arguments{
\item{r}{}
}
\value{
\code{SpatRaster} with values \code{NA} or 1.
}
\description{
Creates a mask where a cell in any layer of \code{r} that is \code{NA} will be returned
as \code{NA}. Similar in intention to \code{mask_all} but will work on larger rasters
because it only holds the values of a single layer in memory at a time, and
returns a mask layer, rather than masking each layer in \code{r}
}
\examples{
r <- example_raster(seed = 10)
s <- example_raster(seed = 11)

r[10:20] <- NA

s[5:15] <- NA

q <- mask_from_all(c(r,s))

library("terra")
plot(c(r,s,q))

}
