% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskpointsdf.R
\name{maskpointsdf}
\alias{maskpointsdf}
\title{Title}
\usage{
maskpointsdf(df, msk)
}
\arguments{
\item{df}{data.frame}

\item{msk}{mask}
}
\value{
tibble of two columsn, lon and lat
}
\description{
Title
}
