% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ref.R
\name{match_ref}
\alias{match_ref}
\title{match ref}
\usage{
match_ref(x, ref, missing_val = NULL, filename = NULL, overwrite = TRUE)
}
\arguments{
\item{x}{\code{SpatRaster} object to bash into shape}

\item{ref}{\code{SpatRaster} reference object}

\item{missing_val}{If \code{NULL} missing value are left as is. Otherwise any
\code{NA} or \code{NaN} values in \code{x} that are not \code{NA} in ref will be replaced by
\code{missing_val}}

\item{filename}{If not \code{NULL} output will be written to filename}

\item{overwrite}{\code{logical}. If \code{filename} not \code{NULL}, then if \code{TRUE},
\code{filename} will be overwritten.}
}
\value{
\code{SpatRaster} of \code{nlyrs(x)} trimmed to extent and resolution of \code{ref}.
}
\description{
Crop, resample, and mask \code{x} against \code{ref}, and optionally
replace any missing values.
}
\examples{
#placeholder example
}
