% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{merge_area}
\alias{merge_area}
\title{Merge rasters over a study area.}
\usage{
merge_area(
  an_area,
  area_source = NULL,
  cell_width = 0,
  cell_height = 0,
  var_names = NULL
)
}
\arguments{
\item{an_area}{A sp object, commonly a shapefile,  or a SDM_area object representing the area of study.}

\item{area_source}{A folder or a raster with variables to merge with.}

\item{cell_width}{The width of cells.}

\item{cell_height}{The height of cells.}

\item{var_names}{A list (or vector) of variable names to keep on cells. Variables area computed using
the average of raster points that over each cell. It try to match each variable name
(ignoring case) in the study area.}
}
\value{
A SpatialPolygonsDataframe containing variables merged with.
}
\description{
Merge rasters over a study area.
}
\examples{
\dontrun{
aaa
}
}
