% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scenario.R
\name{merge_scenario}
\alias{merge_scenario}
\title{Title}
\usage{
merge_scenario(
  an_area = NULL,
  to_merge_scenario = NULL,
  var_names = NULL,
  new_name = NULL,
  dir_path = NULL
)
}
\arguments{
\item{var_names}{}
}
\value{

}
\description{
Title
}
