% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_presence_absence.R
\name{model_data_presence_absence}
\alias{model_data_presence_absence}
\title{Title}
\usage{
model_data_presence_absence(pa_data, covariates)
}
\arguments{
\item{covariates}{}
}
\description{
Title
}
