% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_presence_only.R
\name{model_data_presence_only}
\alias{model_data_presence_only}
\title{Title}
\usage{
model_data_presence_only(presences, absences, covariates)
}
\arguments{
\item{covariates}{}
}
\description{
Title
}
