% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_land.R
\name{nearest_land}
\alias{nearest_land}
\title{Identify nearest land}
\usage{
nearest_land(points, raster, max_distance)
}
\arguments{
\item{points}{anything \code{\link[terra:extract]{terra::extract()}} accepts as the y argument}

\item{raster}{raster}

\item{max_distance}{the map units if raster is projected}
}
\value{
matrix of XY coordinates of nearest cell on mask, or NAs
}
\description{
Adapted from seegSDM
Identify closest neighbouring cell that does not return NA on raster mask
}
