% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_tibble.R
\name{new_long_tibble}
\alias{new_long_tibble}
\title{\code{long_tibble} constructor function}
\usage{
new_long_tibble(x)
}
\arguments{
\item{x}{A \code{tbl_df} object.}
}
\value{
\code{long_tibble}.
}
\description{
\code{long_tibble} constructor function
}
