% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.SDM_area}
\alias{plot.SDM_area}
\title{Plot a SDM_area}
\usage{
\method{plot}{SDM_area}(x, ...)
}
\arguments{
\item{x}{SDM_area object}

\item{...}{Additional parameters}
}
\description{
Plot a SDM_area
}
