% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mpp_rast.R
\name{predict_mpp_rast}
\alias{predict_mpp_rast}
\title{Raster predictions from \code{multispeciesPP}}
\usage{
predict_mpp_rast(
  model,
  data,
  sp,
  type = c("response", "link"),
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{model}{A \code{multispeciesPP} model object}

\item{data}{\code{SpatRaster} object with covariate and bias layers}

\item{sp}{\code{character} Species name to predict}

\item{type}{\code{character}. Prediction scale — \code{"response"} or \code{"link"}.}

\item{filename}{\code{character}to save output}

\item{overwrite}{\code{logical} overwrite existing \code{filename}?}
}
\value{
\code{SpatRaster}
}
\description{
Make spatial predictions from \code{multispeciesPP::predict.multispeciesPP} by passing in data as \code{terra::SpatRaster} layers and returning a \code{SpatRaster}.
}
