% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mpp_rast_all.R
\name{predict_mpp_rast_all}
\alias{predict_mpp_rast_all}
\title{Raster predictions from \code{multispeciesPP} for all species}
\usage{
predict_mpp_rast_all(model, data, type = c("response", "link"))
}
\arguments{
\item{model}{A \code{multispeciesPP} model object}

\item{data}{\code{SpatRaster} object with covariate and bias layers}

\item{type}{\code{character}. Prediction scale — \code{"response"} or \code{"link"}.}
}
\value{
\code{SpatRaster}
}
\description{
Calls \code{sdmtools::predict_mpp_rast} over all species in \code{model}
}
