% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdm.R
\name{predict_sdm}
\alias{predict_sdm}
\title{Predict Species Distribution Model}
\usage{
predict_sdm(
  model,
  covariates,
  type = NULL,
  layer_name = "predicted_distribution"
)
}
\arguments{
\item{model}{A model object.}

\item{covariates}{\code{spatRaster} covariate layers.}

\item{type}{Scale of prediction (response, model, etc.).}

\item{layer_name}{Name for predicted layer.}
}
\value{
\code{spatRaster} prediction from model
}
\description{
Produces raster prediction from SDM based on model and covariate layers.
}
