% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastpointplot.R
\name{rastpointplot}
\alias{rastpointplot}
\title{Raster plot with points}
\usage{
rastpointplot(r, v, pch = 16, cex = 0.5)
}
\arguments{
\item{r}{\code{spatRaster} object}

\item{v}{\code{spatVector} object}

\item{pch}{\code{integer } point symbol. See \code{?par}}

\item{cex}{\code{numeric} point size multiplier. See \code{?par}}
}
\value{
graphical plot
}
\description{
Simple convenience function to plot points over a raster. Useful for quick data checks.
}
