% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gpkg.R
\name{save_gpkg}
\alias{save_gpkg}
\title{Save the vector format geospatial object of the \code{SDM_area} to a geopackage file.}
\usage{
save_gpkg(an_area = NULL, new_name = NULL, dir_path = NULL)
}
\arguments{
\item{an_area}{A \code{SDM_area} object representing the area of the study.}

\item{file_name}{The name of file which the \code{sp} object will be saved. It
can be composed by the file name and the file path. If the file path doesn't exists
it will be created. If the file name parameter is equal to NULL, it is replaced by
a name computed based on attributes of the \code{SDM_area} object.}

\item{file_path}{The path in which the file will be saved. If file_name contains
file path, the file_path parameter is ignored.}
}
\description{
Save the vector format geospatial object of the \code{SDM_area} to a geopackage file.
}
\examples{
\dontrun{
SPDF <- readOGR(
   system.file("vect_files/brasil_uf.gpkg", package="sdmTools"),
   layer = "brasil_uf",
   verbose = F
)

gridded_area <- SPDF \%>\%
 sdm_area("Test area", "EPSG:6933", c(50000, 50000)) \%>\%
 make_grid(var_names = list(), new_name = T)

tmp_dir <- tempdir()

gridded_area \%>\%
   save_gpkg(file_path = tmp_dir)

tmp_dir \%>\%
   paste0("/test_area_grid_50000_epsg_6933.gpkg") \%>\%
   file_info()
}
}
