% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot}
\usage{
save_plot(
  p,
  filename,
  width = 2400,
  height,
  units = c("px", "cm", "mm"),
  res = 300
)
}
\arguments{
\item{p}{A plot object}

\item{filename}{\code{character} to save plot}

\item{width}{In \code{units}}

\item{height}{In \code{units}. If missing, \code{height} will be scaled to \code{width}
at the golden ratio.}

\item{units}{Units of \code{width} and \code{height}. \code{"px"} — pixels, the default,
\code{"cm"}, or \code{"mm"}. \code{"in"} are not allowed on principle.}

\item{res}{resolution, default 300px}
}
\value{
nothing. Writes a plot to \code{filename}
}
\description{
Single line wrapper to save plot as png
}
