% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_predict.R
\name{sdm_predict}
\alias{sdm_predict}
\title{Title}
\usage{
sdm_predict(
  model,
  covariates,
  type = NULL,
  layer_name = "predicted_distribution"
)
}
\arguments{
\item{layer_name}{}
}
\description{
Title
}
