% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tidy.R
\name{sdm_tidy}
\alias{sdm_tidy}
\title{Convert a \code{SDM_area} object to a dataframe.}
\usage{
sdm_tidy(an_area, region = NULL)
}
\arguments{
\item{an_area}{A \code{SDM_area} object representing the area of the study.}

\item{region}{An attribute used to join the primitive shapes to information about study area.
If it is equal to NULL, the information about study area is dropped from the resulting dataframe.}
}
\value{
The \code{SDM_area} object converted to a dataframe. The dataframe is  composed by attributes of the
primitive shapes (polygons or lines) in an object of package \code{sp}
(\url{https://cran.r-project.org/web/packages/sp}) and the information about study area.
}
\description{
Convert a \code{SDM_area} object to a dataframe.
}
\examples{
\dontrun{
SPDF <- rgdal::readOGR(
   system.file("vect_files/brasil_uf.gpkg", package="sdmTools"),
   layer = "brasil_uf",
   verbose = F
)

gridded_area <- SPDF \%>\%
 sdm_area("Test area", "EPSG:6933", c(50000, 50000)) \%>\%
 make_grid(var_names = list(), new_name = T)

a_df <- gridded_area \%>\%
   sdm_tidy(region = "cell_id")

a_df \%>\%
   head(20)
}
}
