% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmtools-package.R
\docType{package}
\name{sdmtools-package}
\alias{sdmtools}
\alias{sdmtools-package}
\title{sdmtools: Utility tools for Species Distribution Modelling}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/idem-lab/sdmtools}
  \item Report bugs at \url{https://github.com/idem-lab/sdmtools/issues}
}

}
\author{
\strong{Maintainer}: Gerry Ryan \email{gerard.ryan@telethonkids.org.au} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID})

Other contributors:
\itemize{
  \item David Duncan [contributor]
  \item Nick Tierney [contributor]
}

}
\keyword{internal}
