% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_layer_names.R
\name{set_layer_names}
\alias{set_layer_names}
\title{set layer names}
\usage{
set_layer_names(x, layernames)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{layernames}{\code{character} of length \code{nlyr(x)}}
}
\value{
\code{SpatRaster}
}
\description{
Convenience function for setting \code{SpatRaster} layer names, enables setting in
piped workflows.
}
\examples{
example_raster() |>
 set_layer_names("Charlie Watts")
}
