% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_levels.R
\name{set_levels}
\alias{set_levels}
\title{set levels}
\usage{
set_levels(x, levs)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{levs}{\code{data.frame}}
}
\value{
\code{SpatRaster}
}
\description{
Convenience function for setting \code{SpatRaster} levels names to enable setting in
piped workflows.
}
\examples{
\dontrun{
categorical_raster |>
 set_levels(
  levs = tribble(
    ~value, ~category,
    30, "URBAN CENTRE",
    23, "DENSE URBAN CLUSTER",
    22, "SEMI-DENSE URBAN CLUSTER",
    21, "SUBURBAN OR PERI-URBAN",
    13, "RURAL CLUSTER",
    12, "LOW DENSITY RURAL",
    11, "VERY LOW DENSITY RURAL",
    10, "WATER"
   )
  ) \%>\%
    as.data.frame()
}

}
