% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_R.R
\name{source_R}
\alias{source_R}
\title{Source all R scripts in a directory}
\usage{
source_R(path = "R", print.names = TRUE)
}
\arguments{
\item{path}{Path of directory; default is  \verb{/R}.}

\item{print.names}{Print path and name of sourced scripts. Default \code{TRUE}.}
}
\value{
Scripts are sourced to local environment. \code{NULL} returned.
}
\description{
Sources all scripts in a specified directory and optionally prints paths.
}
\examples{
#' \dontrun{
  source_R(/Users/frankenstein/project/R)
}
}
