% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_rast.R
\name{split_rast}
\alias{split_rast}
\title{Split rasters}
\usage{
split_rast(x, grain = 4, write_temp = FALSE)
}
\arguments{
\item{x}{A \code{SpatRaster}}

\item{grain}{Grain of splitting.}

\item{write_temp}{}
}
\value{
A list of \code{SpatRaster}s of length \code{grain^2}.
}
\description{
Split rasters
}
\details{
Splits a \code{SpatRaster} up into a \code{grain^2} list of approximately equal
geographic sized rasters covering the extent of  \code{x}
}
\examples{

# Split a raster into four
r <- example_raster()
s <- split_rast(r, grain = 2)
s

# plot with original
 ps <- lapply(
   s,
   FUN = extend,
   y = r
 ) |>
   rast()

 c(r, ps) |>  plot()

}
