% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_rast.R
\name{std_rast}
\alias{std_rast}
\title{Standardise raster}
\usage{
std_rast(x, filename = NULL, overwrite = TRUE)
}
\arguments{
\item{x}{\code{SpatRaster} to standardise}

\item{filename}{Optional \code{character} path and filename to write output}

\item{overwrite}{\code{logical} if \code{TRUE} will overwrite \code{filename}}
}
\description{
Standardises all layers in a \code{SpatRaster} to a scale of 0-1,
by dividing by the maximum value in each layer.
}
\details{
Will break for non-numeric rasters
}
\examples{
example_raster() |>
  std_rast()
}
