% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabble.R
\name{tabble}
\alias{tabble}
\title{Make a single-variable table and display it nicely in console}
\usage{
tabble(x, usena = c("always", "no", "ifany"), n = 50)
}
\arguments{
\item{usena}{per \code{?table}}

\item{n}{print how many rows eh?}
}
\description{
Make a single-variable table and display it nicely in console
}
