% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_bio_names.R
\docType{data}
\name{worldclim_bio_names}
\alias{worldclim_bio_names}
\title{WorldClim \code{"bio"} names}
\format{
\subsection{\code{worldclim} \code{character} vector of names}{
}
}
\source{
\url{https://www.worldclim.org/}
}
\usage{
worldclim_bio_names
}
\value{
\code{character}
}
\description{
Character vector of names for the \code{geodata::worldclim_global}
etc. \code{SpatRaster} layers, as a convenience alternative to the default
detailed and accurate but horridly unreadable format of \code{wc2.1_30s_bio_1}
etc.
}
\keyword{datasets}
