% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writereadrast.R
\name{writereadrast}
\alias{writereadrast}
\title{Write and read back \code{spatRaster}}
\usage{
writereadrast(x, filename, overwrite = TRUE)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}}

\item{filename}{A \code{character} file path and name to save \code{x} to disc.}

\item{overwrite}{\code{logical}; overwrite existing raster.}
}
\value{
A \code{terra::SpatRaster} object reading from disc at \code{filename}.
}
\description{
When a new \code{terra::SpatRaster} is created it is stored in
memory. Writing it to disc and reading the object back from file requires two
steps: \code{terra::writeRaster} then reading and re-assigning via \code{terra::rast}.
\code{writereadrast} wraps these into a single step.
}
\examples{
\dontrun{
# create raster then assign
r <- sdmtools::example_raster()
r <- writereadrast(r, "tootgarook.tif")

# or roll into single step with pipe
q <- sdmtools::example_raster() |>
  writereadrast("frankstonfreeway.tif")
}
}
