% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_gt.R
\name{areas_gt}
\alias{areas_gt}
\title{Drop non-contiguous polygons with area smaller or equal lower_bound.}
\usage{
areas_gt(an_area = NULL, lower_bound = 0, new_name = F)
}
\arguments{
\item{an_area}{A SDM_area object representing the area of study.}

\item{lower_bound}{A lower bound area indicating polygons which it going to dropped out.}

\item{new_name}{A name to new area study after dropping smaller polygons.}
}
\value{
A SDM_area containing a sp object with remaining disaggregated polygons with area greater than lower_bound.
}
\description{
Drop non-contiguous polygons with area smaller or equal lower_bound.
}
\examples{
\dontrun{
new_SPDF <- SP \%>\%
 as("SpatialPolygonsDataFrame")
new_SPDF@data <- list(
 area=c(
   SP[1] \%>\% gArea(),
   SP[2] \%>\% gArea(),
   SP[3] \%>\% gArea()
 )
) \%>\%
 as.data.frame()

plot(new_SPDF)

new_area <- new_SPDF \%>\%
 sdm_area("Removing a single area from a SpatialPolygons study area.", "EPSG:6933", c(50000, 50000)) \%>\%
 areas_gt(0.25, new_name = "Removing a single area from a SpatialPolygons study area.")

plot(new_area$study_area)
}
}
