% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_raster}
\alias{example_raster}
\title{Create example raster for use in examples}
\usage{
example_raster(seed = NULL, layername = NULL)
}
\arguments{
\item{seed}{A seed to pass to \code{set.seed(seed)}}

\item{layername}{\code{character} for layer name passed to \code{names}}
}
\value{
A \code{spatRaster}.
}
\description{
Create example raster for use in examples
}
\examples{
example_raster()

example_raster(
  seed = 3.142,
  layername = "jabberwock_density"
)
}
