% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Extract data from covariate rasters}
\usage{
extract_covariates(
  covariates,
  presences = NULL,
  absences = NULL,
  presences_and_absences = NULL
)
}
\arguments{
\item{covariates}{\code{SpatRaster} object of one or more layers.}

\item{presences}{\code{data.frame} or \code{tibble} of presence locations contaning longitude and latitude as variables \code{x} and \code{y}.}

\item{absences}{\code{data.frame} or \code{tibble} of absence or background locations contaning longitude and latitude as variables \code{x} and \code{y}.}

\item{presences_and_absences}{\code{data.frame} or \code{tibble} of presence presence and absence locations containing longitude and latitude as variables \code{x} and \code{y}.}
}
\value{
A \code{tibble} containing values variables \code{presence} and each of the covariate layers at each location.
}
\description{
Extracts data from raster covariate layers for modelling.
}
\details{
\code{extract_covariates} will run correctly with either \code{presences} only, \code{presences} and \code{absences}, or \code{presences_and_absences} only. (If all three are included it will ignore the last one).
}
\examples{

library(terra)

cov1 <- example_raster(
 seed = -44,
 layername = "cov1"
)
cov2 <- example_raster(
 seed = 15.3,
 layername = "cov2"
)

covs <- c(cov1, cov2)


presences <- example_vector(seed = 68) \%>\%
 as.data.frame(geom = "xy")
absences <- example_vector(seed = 9.6) \%>\%
 as.data.frame(geom = "xy")

extract_covariates(
 covariates = covs,
 presences = presences,
 absences = absences
)

}
\author{
Gerry Ryan
}
