% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_mask.R
\name{inside_mask}
\alias{inside_mask}
\title{title}
\usage{
inside_mask(points, mask)
}
\arguments{
\item{points}{dataframe containing columns named 'longitude' and 'latitude'}

\item{mask}{a raster}
}
\value{
dataframe of longitude and latitude only those rows with
points falling on non-missing pixels. If all points fall on missing
pixels, the function throws an error.
}
\description{
Checks whether longitude and latitude coincide with
non-missing pixels of a raster. The function takes two arguments:
points, a dataframe containing columns named
longitude' and 'latitude', and mask is a raster. Returns a dataframe
of longitude and latitude only those rows with points falling on
non-missing pixels. If all points fall on missing pixels,
the function throws an error.
}
