% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_africa_mask.R
\name{make_africa_mask}
\alias{make_africa_mask}
\title{Make Africa Mask}
\usage{
make_africa_mask(file_name = NULL, res = c("high", "low"))
}
\arguments{
\item{file_name}{Character of file path and name if  mask is to be written to disc.}

\item{res}{Character \code{"high"} or \code{"low"}; correspondiing to resolution of 0.008333333 or 0.04166667 decimal degrees}
}
\value{
\code{spatRaster} mask of Africa; WGS 84 (EPSG:4326).
}
\description{
Makes a \code{spatRaster} mask layer of Africa, based on shapefiles for African nations from the \code{malaraAtlas} package. Function is a little slow, so writing to disc may be preferred.
}
\details{
Created using countries where \code{sdmtools::global_regions$continent == "Africa"} and either the "Explorer__2020_Africa_ITN_Use" or  ""Explorer__2020_walking_only_travel_time_to_healthcare" raster from \code{malariaAtlas::getShp}
}
\examples{
# Create an object in workspace
africa_mask <- make_africa_mask()

# Save to disk
make_africa_mask("africa_mask.tif")

# or do both at once
africa_mask <- make_africa_mask("africa_mask.tif")
}
