% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_africa_mask.R
\name{make_africa_mask}
\alias{make_africa_mask}
\title{Make Africa Mask}
\usage{
make_africa_mask(
  file_name = NULL,
  type = c("raster", "vector"),
  res = c("high", "low"),
  countries = NULL
)
}
\arguments{
\item{file_name}{Character of file path and name if  mask is to be written to disc.}

\item{type}{Character \code{raster} or \code{vector}; to return mask as either \code{SpatRaster} or \code{SpatVector}.}

\item{res}{Character \code{"high"} or \code{"low"}; corresponding to resolution of 0.008333333 or 0.04166667 decimal degrees}

\item{countries}{Character of ISO3 country names. If \code{NULL} returns all countries in Africa.}
}
\value{
\code{SpatRaster} or \code{SpatVector} in WGS 84 (EPSG:4326).
}
\description{
Makes a \code{spatRaster} mask layer of Africa, based on shapefiles for African nations from the \code{malaraAtlas} package. Function is a little slow, so writing to disc may be preferred.
}
\details{
Created using countries where \code{sdmtools::global_regions$continent == "Africa"} from \code{malariaAtlas::getShp} and if \code{type = "raster"} either the "Explorer__2020_Africa_ITN_Use" or  ""Explorer__2020_walking_only_travel_time_to_healthcare" layer.
}
\examples{
# Create an object in workspace
africa_mask <- make_africa_mask()

# Save to disk
make_africa_mask("africa_mask.tif")

# or do both at once
africa_mask <- make_africa_mask("africa_mask.tif")
}
