% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{make_grid}
\alias{make_grid}
\title{Make a grid over study area.}
\usage{
make_grid(an_area = NULL, var_names = NULL, has_centroid = T)
}
\arguments{
\item{an_area}{A SDM_area object representing the area of study.}

\item{var_names}{A list of variable names to keep on cells. Variables area computed using
the average of features (polygons or lines) that over each cell. It try to match each variable name
(ignoring case) in the study area.}

\item{centroid}{A boolean indicating if x_centroid and y_centroid variables must be computed and appended
to variables./}
}
\value{
A SDM_area  object with cells covering the study area. The dataframe contains the variables
matched and computed acoording to each cell.
}
\description{
Make a grid over study area.
}
\examples{
\dontrun{
SPDF <- readOGR(
   system.file("brasil_uf.gpkg", package="sdmTools"),
   layer = "brasil_uf",
   verbose = FALSE
 )
SLDF <- readOGR(
   system.file("hydro_uper_prpy.gpkg", package="sdmTools"),
   layer = "hydro_uper_prpy",
   verbose = FALSE
 )
new_sdm_area <- sdm_area(SPDF, "EPSG:6933", c(50000, 50000)))
gridded_area <- make_grid(
      new_sdm_area,
      var_names = c("Length", "xxx", "Main_ri"),
      centroid = T
 )
gridded_area \%>\% plot()

new_sdm_area <- sdm_area(SLDF, "EPSG:6933", c(50000, 50000)))
gridded_area <- make_grid(
      new_sdm_area,
      var_names = c("Length", "xxx", "Main_ri"),
      centroid = T
 )
gridded_area \%>\% plot()
}
}
