% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_area.R
\name{merge_area}
\alias{merge_area}
\title{Merge rasters over a gridded study area.}
\usage{
merge_area(
  an_area = NULL,
  to_merge_area = NULL,
  var_names = NULL,
  new_name = F
)
}
\arguments{
\item{an_area}{A SDM_area object with cells covering the study area.}

\item{var_names}{A list of variable names to keep on cells. Variables area computed using
the average of raster points that over each cell. It try to match each variable name
(ignoring case) in the study area.}

\item{area_source}{A path to a folder or a Raster* object with variables to merge with.}
}
\value{
A SDM_area_gridded object containing variables merged with. If the CRS of the Raster* is
different from the CRS of the SDM_area object, it is reproject. The merging process
produces cells intersecting SDM_area object and Raster*.
}
\description{
Merge rasters over a gridded study area.
}
\examples{
\dontrun{
aaa
}
}
