% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mpp_rast_all.R
\name{predict_mpp_rast_all}
\alias{predict_mpp_rast_all}
\title{Raster predictions from \code{multispeciesPP} for all species}
\usage{
predict_mpp_rast_all(
  model,
  data,
  type = c("response", "link"),
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{model}{A \code{multispeciesPP} model object}

\item{data}{\code{SpatRaster} object with covariate and bias layers}

\item{type}{\code{character}. Prediction scale — \code{"response"} or \code{"link"}.}

\item{filename}{\code{character}to save output}

\item{overwrite}{\code{logical} overwrite existing \code{filename}?}
}
\value{
\code{SpatRaster}
}
\description{
Calls \code{sdmtools::predict_mpp_rast} over all species in \code{model}
}
