% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{sdm_area}
\alias{sdm_area}
\title{Creates a Study Area}
\usage{
sdm_area(an_area = NULL)
}
\arguments{
\item{an_area}{A sp object, commonly a shapefile,  or a SDM_area object representing the area of study.}
}
\value{
An object representing a study area containing a sp object. If polygons are invalid because topological
erros they are corrected. Furthermore, if no coordinate reference system are used it will be set to EPSG:4326.
}
\description{
Creates a Study Area
}
\examples{
\dontrun{
main <-  cbind(
c(0, 0, 1, 1),
c(0, 1, 1, 0)
)
secondary <- cbind(
  c(1, 1.3, 1.3, 1),
  c(1, 1.0, 0.7, 0.7)
)
hole <- main/3 + 1/3
island = cbind(
  c(1.05, 1.05, 1.55, 1.55),
  c(0, .5, .5, 0)
)

P <- Polygons(
  ID = 1,
  list(
    Polygon(main),
    Polygon(hole, hole = TRUE),
    Polygon(island),
    Polygon(secondary)
  )
)

SP = SpatialPolygons(list(P))
crs(SP) <- CRS("EPSG:6933")

new_sdm_area <- sdm_area(SP)

class(new_sdm_area)
}
}
