% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_area.R
\name{sdm_area}
\alias{sdm_area}
\title{Creates a Study Area}
\usage{
sdm_area(an_area = NULL, name = NULL, epsg_code = NULL, a_res = NULL)
}
\arguments{
\item{an_area}{An object of package \code{sp} (\url{https://cran.r-project.org/web/packages/sp})
or a path to a file, commonly a shapefile or geopackage, representing the area of study.}

\item{name}{A name do describe the study area.}

\item{epsg_code}{A valid EPSG code, for example EPSG:4326.}

\item{a_res}{A vector containing the resolution of the study area. The format is two numeric values
(width and height) according to \code{epsg_code} used. So, the numeric values can express different
types of measurement units, for example deegres or meters.}
}
\value{
An object representing a study area containing a \code{sp} object.

Occasional topological errors on polygons of the object are corrected. As a side effect holes inside
polygons are removed.

If the \code{epsg_code} of the study area is different from the epsg_code passed to the function, the study
area is reproject.
}
\description{
A study area (\code{SDm_area}) represents a geographic location delimited by primitive shapes (lines and polygons)  and
associated attribute information about characteristics of that location.
}
\details{
The attribute information of the study area are constant in time. In contrast, the \code{SDM_area} object
has an array attribute storing multiple scenarios, representing attribute information about characteristics of the
location that change in time.
}
\examples{
\dontrun{
SPDF <- readOGR(
   system.file("brasil_uf.gpkg", package="sdmTools"),
   layer = "brasil_uf",
   verbose = FALSE
)
new_sdm_area <- sdm_area(SPDF, "Brasil", "EPSG:6933", c(50000, 50000)))

class(new_sdm_area)

plot(new_sdm_area)
}

}
