\name{SDMTools-package}
\alias{SDMTools-package}
\alias{SDMTools}
\docType{package}
\title{
Species Distribution Modelling Tools:
Tools for processing data associated with species distribution modelling exercises.
}
\description{
This is a set of tools that are useful for applications of species distribution 
modelling.
}
\details{
\tabular{ll}{
Package: \tab SDMTools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-05-05\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jeremy VanDerWal, Luke Shoo and Stephanie Januchowski. This package was made possible in part by support from Lorena Falconi and Collin Storlie, and financial support from the Australian Research Council & ARC Research Network for Earth System Science.

Maintainer: Jeremy VanDerWal \email{jjvanderwal@gmail.com}
}

\keyword{ package }

