\name{SDMTools-package}
\Rdversion{1.1}
\alias{SDMTools-package}
\alias{SDMTools}
\docType{package}
\title{
Tools for processing data associated with species distribution modelling exercises.
}
\description{
This is a set of tools that are useful for applications of species distribution 
modelling.
}
\details{
\tabular{ll}{
Package: \tab SDMTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-11-20\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jeremy VanDerWal, Luke Shoo and Stephanie Januchowski

Maintainer: Jeremy VanDerWal \email{jjvanderwal@gmail.com}
}

\keyword{ package }

