% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{package}
\name{sdmtools-package}
\alias{sdmtools}
\alias{sdmtools-package}
\title{sdmtools: Species Distribution Modelling Tools: Tools for processing data associated with species distribution modelling exercises}
\description{
This packages provides a set of tools for post processing the outcomes of species distribution modeling exercises. It includes novel methods for comparing models and tracking changes in distributions through time. It further includes methods for visualizing outcomes, selecting thresholds, calculating measures of accuracy and landscape fragmentation statistics, etc.. This package was made possible in part by financial support from the Australian Research Council & ARC Research Network for Earth System Science.
}
\author{
\strong{Maintainer}: Jeremy VanDerWal \email{jjvanderwal@gmail.com}

Authors:
\itemize{
  \item Lorena Falconi
  \item Stephanie Januchowski
  \item Luke Shoo
  \item Collin Storlie
}

Other contributors:
\itemize{
  \item Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [contributor]
}

}
\keyword{internal}
