% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmTools-package.R
\docType{package}
\name{sdmTools-package}
\alias{sdmTools}
\alias{sdmTools-package}
\title{sdmTools: It provide tools to help in Species Distribution Modelling}
\description{
It provide tools to help in Species Distribution Modelling.
}
\author{
\strong{Maintainer}: Reginaldo Ré \email{reginaldo.re@gmail.com} (\href{https://orcid.org/0000-0001-6452-3466}{ORCID})

}
\keyword{internal}
