% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_rast.R
\name{standardise_rast}
\alias{standardise_rast}
\title{Standardise Raster}
\usage{
standardise_rast(x)
}
\arguments{
\item{x}{A \code{SpatRaster} object.}
}
\value{
A \code{SpatRaster} objects with mean of zero and standard deviation of one.
}
\description{
Standardises \code{SpatRaster} objects to mean of zero and standard deviation of one.
}
\examples{
r <- example_raster()

standardise_rast(r)

}
